move setchannelsmixin to meas_info
