+++ mne/evoked.py
@@ -12,7 +12,7 @@ from copy import deepcopy
 import numpy as np
 
 from .baseline import rescale, _log_rescale, _check_baseline
-from .channels.channels import UpdateChannelsMixin, SetChannelsMixin, InterpolationMixin
+from .channels.channels import UpdateChannelsMixin, InterpolationMixin
 from .channels.layout import _merge_ch_data, _pair_grad_sensors
 from .defaults import _INTERPOLATION_DEFAULT, _EXTRAPOLATE_DEFAULT, _BORDER_DEFAULT
 from .filter import detrend, FilterMixin, _check_fun
@@ -55,6 +55,7 @@ from .io.tree import dir_tree_find
 from .io.pick import pick_types, _picks_to_idx, _FNIRS_CH_TYPES_SPLIT
 from .io.meas_info import (
     ContainsMixin,
+    SetChannelsMixin,
     read_meas_info,
     write_meas_info,
     _read_extended_ch_info,
