+++ mne/io/meas_info.py
@@ -22,9 +22,10 @@ from .pick import (
     _get_channel_types,
     get_channel_type_constants,
     pick_types,
+    _picks_to_idx,
     _contains_ch_type,
 )
-from .constants import FIFF, _coord_frame_named
+from .constants import FIFF, _coord_frame_named, _ch_unit_mul_named
 from .open import fiff_open
 from .tree import dir_tree_find
 from .tag import (
@@ -421,6 +422,465 @@ class MontageMixin:
 
         info = self if isinstance(self, Info) else self.info
         _set_montage(info, montage, match_case, match_alias, on_missing)
+        return self
+
+
+channel_type_constants = get_channel_type_constants(include_defaults=True)
+_human2fiff = {
+    k: v.get("kind", FIFF.FIFFV_COIL_NONE) for k, v in channel_type_constants.items()
+}
+_human2unit = {
+    k: v.get("unit", FIFF.FIFF_UNIT_NONE) for k, v in channel_type_constants.items()
+}
+_unit2human = {
+    FIFF.FIFF_UNIT_V: "V",
+    FIFF.FIFF_UNIT_T: "T",
+    FIFF.FIFF_UNIT_T_M: "T/m",
+    FIFF.FIFF_UNIT_MOL: "M",
+    FIFF.FIFF_UNIT_NONE: "NA",
+    FIFF.FIFF_UNIT_CEL: "C",
+    FIFF.FIFF_UNIT_S: "S",
+    FIFF.FIFF_UNIT_PX: "px",
+}
+
+
+def _check_set(ch, projs, ch_type):
+    """Ensure type change is compatible with projectors."""
+    new_kind = _human2fiff[ch_type]
+    if ch["kind"] != new_kind:
+        for proj in projs:
+            if ch["ch_name"] in proj["data"]["col_names"]:
+                raise RuntimeError(
+                    "Cannot change channel type for channel %s "
+                    'in projector "%s"' % (ch["ch_name"], proj["desc"])
+                )
+    ch["kind"] = new_kind
+
+
+class SetChannelsMixin(MontageMixin):
+    """Mixin class for Raw, Evoked, Epochs."""
+
+    @verbose
+    def set_eeg_reference(
+        self,
+        ref_channels="average",
+        projection=False,
+        ch_type="auto",
+        forward=None,
+        *,
+        joint=False,
+        verbose=None,
+    ):
+        """Specify which reference to use for EEG data.
+
+        Use this function to explicitly specify the desired reference for EEG.
+        This can be either an existing electrode or a new virtual channel.
+        This function will re-reference the data according to the desired
+        reference.
+
+        Parameters
+        ----------
+        %(ref_channels_set_eeg_reference)s
+        %(projection_set_eeg_reference)s
+        %(ch_type_set_eeg_reference)s
+        %(forward_set_eeg_reference)s
+        %(joint_set_eeg_reference)s
+        %(verbose)s
+
+        Returns
+        -------
+        inst : instance of Raw | Epochs | Evoked
+            Data with EEG channels re-referenced. If ``ref_channels='average'``
+            and ``projection=True`` a projection will be added instead of
+            directly re-referencing the data.
+        %(set_eeg_reference_see_also_notes)s
+        """
+        from ..io.reference import set_eeg_reference
+
+        return set_eeg_reference(
+            self,
+            ref_channels=ref_channels,
+            copy=False,
+            projection=projection,
+            ch_type=ch_type,
+            forward=forward,
+            joint=joint,
+        )[0]
+
+    def _get_channel_positions(self, picks=None):
+        """Get channel locations from info.
+
+        Parameters
+        ----------
+        picks : str | list | slice | None
+            None gets good data indices.
+
+        Notes
+        -----
+        .. versionadded:: 0.9.0
+        """
+        picks = _picks_to_idx(self.info, picks)
+        chs = self.info["chs"]
+        pos = np.array([chs[k]["loc"][:3] for k in picks])
+        n_zero = np.sum(np.sum(np.abs(pos), axis=1) == 0)
+        if n_zero > 1:  # XXX some systems have origin (0, 0, 0)
+            raise ValueError(
+                "Could not extract channel positions for " "{} channels".format(n_zero)
+            )
+        return pos
+
+    def _set_channel_positions(self, pos, names):
+        """Update channel locations in info.
+
+        Parameters
+        ----------
+        pos : array-like | np.ndarray, shape (n_points, 3)
+            The channel positions to be set.
+        names : list of str
+            The names of the channels to be set.
+
+        Notes
+        -----
+        .. versionadded:: 0.9.0
+        """
+        if len(pos) != len(names):
+            raise ValueError(
+                "Number of channel positions not equal to " "the number of names given."
+            )
+        pos = np.asarray(pos, dtype=np.float64)
+        if pos.shape[-1] != 3 or pos.ndim != 2:
+            msg = "Channel positions must have the shape (n_points, 3) " "not %s." % (
+                pos.shape,
+            )
+            raise ValueError(msg)
+        for name, p in zip(names, pos):
+            if name in self.ch_names:
+                idx = self.ch_names.index(name)
+                self.info["chs"][idx]["loc"][:3] = p
+            else:
+                msg = "%s was not found in the info. Cannot be updated." % name
+                raise ValueError(msg)
+
+    @verbose
+    def set_channel_types(self, mapping, *, on_unit_change="warn", verbose=None):
+        """Specify the sensor types of channels.
+
+        Parameters
+        ----------
+        mapping : dict
+            A dictionary mapping channel names to sensor types, e.g.,
+            ``{'EEG061': 'eog'}``.
+        on_unit_change : ``'raise'`` | ``'warn'`` | ``'ignore'``
+            What to do if the measurement unit of a channel is changed
+            automatically to match the new sensor type.
+
+            .. versionadded:: 1.4
+        %(verbose)s
+
+        Returns
+        -------
+        inst : instance of Raw | Epochs | Evoked
+            The instance (modified in place).
+
+            .. versionchanged:: 0.20
+               Return the instance.
+
+        Notes
+        -----
+        The following sensor types are accepted:
+
+            ecg, eeg, emg, eog, exci, ias, misc, resp, seeg, dbs, stim, syst,
+            ecog, hbo, hbr, fnirs_cw_amplitude, fnirs_fd_ac_amplitude,
+            fnirs_fd_phase, fnirs_od, eyetrack_pos, eyetrack_pupil,
+            temperature, gsr
+
+        .. versionadded:: 0.9.0
+        """
+        ch_names = self.info["ch_names"]
+
+        # first check and assemble clean mappings of index and name
+        unit_changes = dict()
+        for ch_name, ch_type in mapping.items():
+            if ch_name not in ch_names:
+                raise ValueError(
+                    "This channel name (%s) doesn't exist in " "info." % ch_name
+                )
+
+            c_ind = ch_names.index(ch_name)
+            if ch_type not in _human2fiff:
+                raise ValueError(
+                    "This function cannot change to this "
+                    "channel type: %s. Accepted channel types "
+                    "are %s." % (ch_type, ", ".join(sorted(_human2unit.keys())))
+                )
+            # Set sensor type
+            _check_set(self.info["chs"][c_ind], self.info["projs"], ch_type)
+            unit_old = self.info["chs"][c_ind]["unit"]
+            unit_new = _human2unit[ch_type]
+            if unit_old not in _unit2human:
+                raise ValueError(
+                    "Channel '%s' has unknown unit (%s). Please "
+                    "fix the measurement info of your data." % (ch_name, unit_old)
+                )
+            if unit_old != _human2unit[ch_type]:
+                this_change = (_unit2human[unit_old], _unit2human[unit_new])
+                if this_change not in unit_changes:
+                    unit_changes[this_change] = list()
+                unit_changes[this_change].append(ch_name)
+                # reset unit multiplication factor since the unit has now changed
+                self.info["chs"][c_ind]["unit_mul"] = _ch_unit_mul_named[0]
+            self.info["chs"][c_ind]["unit"] = _human2unit[ch_type]
+            if ch_type in ["eeg", "seeg", "ecog", "dbs"]:
+                coil_type = FIFF.FIFFV_COIL_EEG
+            elif ch_type == "hbo":
+                coil_type = FIFF.FIFFV_COIL_FNIRS_HBO
+            elif ch_type == "hbr":
+                coil_type = FIFF.FIFFV_COIL_FNIRS_HBR
+            elif ch_type == "fnirs_cw_amplitude":
+                coil_type = FIFF.FIFFV_COIL_FNIRS_CW_AMPLITUDE
+            elif ch_type == "fnirs_fd_ac_amplitude":
+                coil_type = FIFF.FIFFV_COIL_FNIRS_FD_AC_AMPLITUDE
+            elif ch_type == "fnirs_fd_phase":
+                coil_type = FIFF.FIFFV_COIL_FNIRS_FD_PHASE
+            elif ch_type == "fnirs_od":
+                coil_type = FIFF.FIFFV_COIL_FNIRS_OD
+            elif ch_type == "eyetrack_pos":
+                coil_type = FIFF.FIFFV_COIL_EYETRACK_POS
+            elif ch_type == "eyetrack_pupil":
+                coil_type = FIFF.FIFFV_COIL_EYETRACK_PUPIL
+            else:
+                coil_type = FIFF.FIFFV_COIL_NONE
+            self.info["chs"][c_ind]["coil_type"] = coil_type
+
+        msg = "The unit for channel(s) {0} has changed from {1} to {2}."
+        for this_change, names in unit_changes.items():
+            _on_missing(
+                on_missing=on_unit_change,
+                msg=msg.format(", ".join(sorted(names)), *this_change),
+                name="on_unit_change",
+            )
+
+        return self
+
+    @verbose
+    def rename_channels(self, mapping, allow_duplicates=False, *, verbose=None):
+        """Rename channels.
+
+        Parameters
+        ----------
+        %(mapping_rename_channels_duplicates)s
+        %(verbose)s
+
+        Returns
+        -------
+        inst : instance of Raw | Epochs | Evoked
+            The instance (modified in place).
+
+            .. versionchanged:: 0.20
+               Return the instance.
+
+        Notes
+        -----
+        .. versionadded:: 0.9.0
+        """
+        from . import BaseRaw
+        from ..channels.channels import rename_channels
+
+        ch_names_orig = list(self.info["ch_names"])
+        rename_channels(self.info, mapping, allow_duplicates)
+
+        # Update self._orig_units for Raw
+        if isinstance(self, BaseRaw):
+            # whatever mapping was provided, now we can just use a dict
+            mapping = dict(zip(ch_names_orig, self.info["ch_names"]))
+            for old_name, new_name in mapping.items():
+                if old_name in self._orig_units:
+                    self._orig_units[new_name] = self._orig_units.pop(old_name)
+            ch_names = self.annotations.ch_names
+            for ci, ch in enumerate(ch_names):
+                ch_names[ci] = tuple(mapping.get(name, name) for name in ch)
+
+        return self
+
+    @verbose
+    def plot_sensors(
+        self,
+        kind="topomap",
+        ch_type=None,
+        title=None,
+        show_names=False,
+        ch_groups=None,
+        to_sphere=True,
+        axes=None,
+        block=False,
+        show=True,
+        sphere=None,
+        *,
+        verbose=None,
+    ):
+        """Plot sensor positions.
+
+        Parameters
+        ----------
+        kind : str
+            Whether to plot the sensors as 3d, topomap or as an interactive
+            sensor selection dialog. Available options 'topomap', '3d',
+            'select'. If 'select', a set of channels can be selected
+            interactively by using lasso selector or clicking while holding
+            control key. The selected channels are returned along with the
+            figure instance. Defaults to 'topomap'.
+        ch_type : None | str
+            The channel type to plot. Available options ``'mag'``, ``'grad'``,
+            ``'eeg'``, ``'seeg'``, ``'dbs'``, ``'ecog'``, ``'all'``. If ``'all'``, all
+            the available mag, grad, eeg, seeg, dbs, and ecog channels are plotted. If
+            None (default), then channels are chosen in the order given above.
+        title : str | None
+            Title for the figure. If None (default), equals to ``'Sensor
+            positions (%%s)' %% ch_type``.
+        show_names : bool | array of str
+            Whether to display all channel names. If an array, only the channel
+            names in the array are shown. Defaults to False.
+        ch_groups : 'position' | array of shape (n_ch_groups, n_picks) | None
+            Channel groups for coloring the sensors. If None (default), default
+            coloring scheme is used. If 'position', the sensors are divided
+            into 8 regions. See ``order`` kwarg of :func:`mne.viz.plot_raw`. If
+            array, the channels are divided by picks given in the array.
+
+            .. versionadded:: 0.13.0
+        to_sphere : bool
+            Whether to project the 3d locations to a sphere. When False, the
+            sensor array appears similar as to looking downwards straight above
+            the subject's head. Has no effect when kind='3d'. Defaults to True.
+
+            .. versionadded:: 0.14.0
+        axes : instance of Axes | instance of Axes3D | None
+            Axes to draw the sensors to. If ``kind='3d'``, axes must be an
+            instance of Axes3D. If None (default), a new axes will be created.
+
+            .. versionadded:: 0.13.0
+        block : bool
+            Whether to halt program execution until the figure is closed.
+            Defaults to False.
+
+            .. versionadded:: 0.13.0
+        show : bool
+            Show figure if True. Defaults to True.
+        %(sphere_topomap_auto)s
+        %(verbose)s
+
+        Returns
+        -------
+        fig : instance of Figure
+            Figure containing the sensor topography.
+        selection : list
+            A list of selected channels. Only returned if ``kind=='select'``.
+
+        See Also
+        --------
+        mne.viz.plot_layout
+
+        Notes
+        -----
+        This function plots the sensor locations from the info structure using
+        matplotlib. For drawing the sensors using PyVista see
+        :func:`mne.viz.plot_alignment`.
+
+        .. versionadded:: 0.12.0
+        """
+        from ..viz.utils import plot_sensors
+
+        return plot_sensors(
+            self.info,
+            kind=kind,
+            ch_type=ch_type,
+            title=title,
+            show_names=show_names,
+            ch_groups=ch_groups,
+            to_sphere=to_sphere,
+            axes=axes,
+            block=block,
+            show=show,
+            sphere=sphere,
+            verbose=verbose,
+        )
+
+    @verbose
+    def anonymize(self, daysback=None, keep_his=False, verbose=None):
+        """Anonymize measurement information in place.
+
+        Parameters
+        ----------
+        %(daysback_anonymize_info)s
+        %(keep_his_anonymize_info)s
+        %(verbose)s
+
+        Returns
+        -------
+        inst : instance of Raw | Epochs | Evoked
+            The modified instance.
+
+        Notes
+        -----
+        %(anonymize_info_notes)s
+
+        .. versionadded:: 0.13.0
+        """
+        anonymize_info(self.info, daysback=daysback, keep_his=keep_his, verbose=verbose)
+        self.set_meas_date(self.info["meas_date"])  # unify annot update
+        return self
+
+    def set_meas_date(self, meas_date):
+        """Set the measurement start date.
+
+        Parameters
+        ----------
+        meas_date : datetime | float | tuple | None
+            The new measurement date.
+            If datetime object, it must be timezone-aware and in UTC.
+            A tuple of (seconds, microseconds) or float (alias for
+            ``(meas_date, 0)``) can also be passed and a datetime
+            object will be automatically created. If None, will remove
+            the time reference.
+
+        Returns
+        -------
+        inst : instance of Raw | Epochs | Evoked
+            The modified raw instance. Operates in place.
+
+        See Also
+        --------
+        mne.io.Raw.anonymize
+
+        Notes
+        -----
+        If you want to remove all time references in the file, call
+        :func:`mne.io.anonymize_info(inst.info) <mne.io.anonymize_info>`
+        after calling ``inst.set_meas_date(None)``.
+
+        .. versionadded:: 0.20
+        """
+        from ..annotations import _handle_meas_date
+
+        meas_date = _handle_meas_date(meas_date)
+        with self.info._unlock():
+            self.info["meas_date"] = meas_date
+
+        # clear file_id and meas_id if needed
+        if meas_date is None:
+            for key in ("file_id", "meas_id"):
+                value = self.info.get(key)
+                if value is not None:
+                    assert "msecs" not in value
+                    value["secs"] = DATE_NONE[0]
+                    value["usecs"] = DATE_NONE[1]
+                    # The following copy is needed for a test CTF dataset
+                    # otherwise value['machid'][:] = 0 would suffice
+                    _tmp = value["machid"].copy()
+                    _tmp[:] = 0
+                    value["machid"] = _tmp
+
+        if hasattr(self, "annotations"):
+            self.annotations._orig_time = meas_date
         return self
